Updated useColorGeneration hook.
