+++ src/components/ColorBarList.tsx
@@ -1,16 +1,7 @@
 import CompareAll from './CompareAll';
 import ThirdColor from './ThirdColor';
-import { useState } from 'react';
 import Neighbor from './Neighbor';
-
-const defaultColors = [
-    '#6975ff',
-    '#084fd7',
-    '#64bdc6',
-    '#eeca34',
-    '#fe7135',
-    '#fd2a2a',
-];
+import useColorGeneration from '../hooks/useColorGeneration';
 
 interface ColorBarListProps {
     selectedContrast: number;
@@ -21,29 +12,12 @@ const ColorBarList = ({
     selectedContrast,
     selectedMode,
 }: ColorBarListProps): JSX.Element => {
-    const [colorBars, setColorBars] = useState<string[]>(defaultColors);
-
-    const addColorBar = (): void => {
-        if (colorBars.length < 10) {
-            const randomIndex = Math.floor(
-                Math.random() * defaultColors.length,
-            );
-            const newColor = defaultColors[randomIndex];
-            setColorBars([...colorBars, newColor]);
-        }
-    };
-
-    const removeColorBar = (index: number): void => {
-        if (colorBars.length > 2) {
-            setColorBars(colorBars.filter((_, i) => i !== index));
-        }
-    };
-
-    const handleColorChange = (index: number, newColor: string): void => {
-        const updatedColors = [...colorBars];
-        updatedColors[index] = newColor;
-        setColorBars(updatedColors);
-    };
+    const {
+        colors: colorBars,
+        handleColorChange,
+        addColorBar,
+        removeColorBar,
+    } = useColorGeneration();
 
     if (selectedMode === 'third') {
         return <ThirdColor selectedContrast={selectedContrast} />;
@@ -64,7 +38,9 @@ const ColorBarList = ({
                 colorBars={colorBars}
                 selectedMode={selectedMode}
                 selectedContrast={selectedContrast}
-                setColorBars={setColorBars}
+                setColorBars={(setColors) => {
+                    handleColorChange(0, setColors[0]);
+                }}
                 handleColorChange={handleColorChange}
                 removeColorBar={removeColorBar}
                 addColorBar={addColorBar}
